refactor: BaseComposeActivity 추가
