+++ app/src/main/java/nextstep/shoppingcart/ui/base/BaseComposeActivity.kt
@@ -1,0 +1,31 @@
+package nextstep.shoppingcart.ui.base
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
+
+
+abstract class BaseComposeActivity : ComponentActivity() {
+    abstract val content: @Composable () -> Unit
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            ShoppingCartTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    content()
+                }
+            }
+        }
+    }
+}
