+++ app/src/main/java/nextstep/shoppingcart/ui/view/product/list/ProductListActivity.kt
@@ -1,27 +1,11 @@
 package nextstep.shoppingcart.ui.view.product.list
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
+import androidx.compose.runtime.Composable
+import nextstep.shoppingcart.ui.base.BaseComposeActivity
 
-class ProductListActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            ShoppingCartTheme {
-                // A surface container using the 'background' color from the theme
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    ProductListScreen()
-                }
-            }
-        }
+class ProductListActivity : BaseComposeActivity() {
+
+    override val content: @Composable () -> Unit = {
+        ProductListScreen()
     }
 }
